README.md
Here is a simple calendar that will allow the user to input tasks and schedule out their work day. Tasks that are in the future will appear green, while tasks that are in the present will appear red, and when tasks are in the past they will appear gray. The user can save their input by clicking the save button at the end of each row. 

Enjoy!